requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.4.0; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
futures==3.0.5
pika==1.3.2; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.11'
